requirements.txt
langchain
langchain-community
langchain-openai
langchain-text-splitters
langchain-classic
langchain-core
faiss-cpu
pypdf
python-dotenv
streamlit
